edit page added and completed

+++ app/Http/Controllers/fundController.php
@@ -2,6 +2,7 @@
 
 namespace App\Http\Controllers;
 
+use App\country;
 use App\field;
 use App\fund;
 use App\organization;
@@ -29,12 +30,25 @@ class fundController extends Controller
 
 
     public function show($id){
-        $funds = fund::find($id);
-        $categories = $funds->tags->all();
-        $organizations = $funds->organization;
-        $country = $organizations->country;
-        $fields = $funds->fields->all();
-        return view('fundShow')->with(compact('funds', 'categories', 'organizations', 'country', 'fields'));
+        $funds = $this->getFunds();
+        $countries = $this->getCountries();
+        $fields = $this->getFields();
+        $categories = $this->getCategories();
+        $organizations = $this->getOrganizations();
+        $count = [];
+        $i = 1;
+        foreach ($funds as $f){
+            array_push($count, $i);
+            $i++;
+        }
+
+
+        $selectedfund = fund::find($id);
+        $fundCategories = $selectedfund->tags->all();
+        $fundOrganizations = $selectedfund->organization;
+        $fundCountries = $fundOrganizations->country;
+        $fundFields = $selectedfund->fields->all();
+        return view('fundEdit')->with(compact('funds', 'categories', 'organizations', 'countries', 'fields', 'selectedfund', 'fundCategories', 'fundCountries', 'fundFields', 'fundOrganizations'));
     }
 
     public function update(Request $r, $id){
@@ -44,9 +58,12 @@ class fundController extends Controller
         $fund = fund::find($fund_id);
 
         if($field_name!= "organization" && $field_name!="field" && $field_name!= "category"){
+            if($field_name == 'name' || $field_name == 'rating')
+                if(!$field_value)
+                    return 'no';
             $fund->update([$field_name => $field_value]);
 //            return a part of view we need to reload
-            return redirect('/fund/1');
+            return redirect('/fund/'.$fund_id);
         }
 
         if($field_name == "organization"){
@@ -54,16 +71,52 @@ class fundController extends Controller
             $fund->organization()->dissociate();
             $fund->organization()->associate(organization::find($field_value));
         } elseif ($field_name == "field"){
+            if(!$field_value)
+                return 'no';
             $fund->fields()->detach();
             foreach ($field_value as $field)
                 $fund->fields()->attach(field::find($field));
         } elseif ($field_name == "category") {
+            if(!$field_value)
+                return 'no';
             $fund->tags()->detach();
             foreach ($field_value as $category)
                 $fund->tags()->attach(tag::find($category));
         }
         // return the part of page needed to reload
         $fund->save();
-        return redirect('/fund/1');
+        return redirect('/fund/'.$fund_id);
     }
+
+
+    public function del(Request $r, $id){
+        $fund_id = $id;
+        $fund = fund::find($fund_id);
+        $fund->organization()->dissociate();
+        $fund->fields()->detach();
+        $fund->tags()->detach();
+        $fund->delete();
+
+    }
+
+    private function getFunds(){
+        return fund::all()->take(10);
+    }
+
+    private function getCategories(){
+        return tag::all();
+    }
+
+    private function getCountries(){
+        return country::all();
+    }
+
+    private function getOrganizations(){
+        return organization::with('country')->get();
+    }
+
+    private function getFields(){
+        return field::all();
+    }
+
 }
